Moved actor implementation to unamanged actor
