+++ src/main/java/io/dropwizard/actors/actor/BaseActor.java
@@ -1,20 +1,17 @@
 package io.dropwizard.actors.actor;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Lists;
-import com.rabbitmq.client.*;
+import com.rabbitmq.client.AMQP;
+import com.rabbitmq.client.MessageProperties;
 import io.dropwizard.actors.connectivity.RMQConnection;
-import io.dropwizard.actors.retry.RetryStrategy;
 import io.dropwizard.actors.retry.RetryStrategyFactory;
 import io.dropwizard.lifecycle.Managed;
-import lombok.*;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.ToString;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.ClassUtils;
 
-import java.io.IOException;
-import java.util.Collections;
-import java.util.List;
 import java.util.Set;
 
 /**
@@ -26,18 +23,8 @@ import java.util.Set;
 @Slf4j
 public abstract class BaseActor<Message> implements Managed {
 
-    private final String name;
-    private final ActorConfig config;
-    private final RMQConnection connection;
-    private final ObjectMapper mapper;
-    private final Class<? extends Message> clazz;
+    private final UnmanagedBaseActor<Message> actorImpl;
     private final Set<Class<?>> droppedExceptionTypes;
-    private final int prefetchCount;
-    private final String queueName;
-    private final RetryStrategy retryStrategy;
-
-    private Channel publishChannel;
-    private List<Handler> handlers = Lists.newArrayList();
 
     protected BaseActor(
             String name,
@@ -47,15 +34,8 @@ public abstract class BaseActor<Message> implements Managed {
             RetryStrategyFactory retryStrategyFactory,
             Class<? extends Message> clazz,
             Set<Class<?>> droppedExceptionTypes) {
-        this.name = name;
-        this.config = config;
-        this.connection = connection;
-        this.mapper = mapper;
-        this.clazz = clazz;
         this.droppedExceptionTypes = droppedExceptionTypes;
-        this.prefetchCount = config.getPrefetchCount();
-        this.queueName  = String.format("%s.%s", config.getPrefix(), name);
-        this.retryStrategy = retryStrategyFactory.create(config.getRetryConfig());
+        actorImpl = new UnmanagedBaseActor<>(name, config, connection, mapper, retryStrategyFactory, clazz, this::handle, this::isExceptionIgnorable);
     }
 
     abstract protected boolean handle(Message message) throws Exception;
@@ -66,77 +46,8 @@ public abstract class BaseActor<Message> implements Managed {
                 .anyMatch(exceptionType -> ClassUtils.isAssignable(t.getClass(), exceptionType));
     }
 
-    private class Handler extends DefaultConsumer {
-        private final ObjectMapper mapper;
-        private final Class<? extends Message> clazz;
-        private final Set<Class<?>> droppedExceptionTypes;
-        private final BaseActor<Message> actor;
-
-        @Getter
-        @Setter
-        private String tag;
-
-        private Handler(Channel channel,
-                        ObjectMapper mapper,
-                        Class<? extends Message> clazz,
-                        Set<Class<?>> droppedExceptionTypes,
-                        int prefetchCount,
-                        BaseActor<Message> actor) throws Exception {
-            super(channel);
-            this.mapper = mapper;
-            this.clazz = clazz;
-            this.droppedExceptionTypes = droppedExceptionTypes;
-            this.actor = actor;
-            getChannel().basicQos(prefetchCount);
-        }
-
-        @Override
-        public void handleDelivery(String consumerTag, Envelope envelope,
-                                   AMQP.BasicProperties properties, byte[] body) throws IOException {
-            try {
-                final Message message = mapper.readValue(body, clazz);
-
-                boolean success = retryStrategy.execute(() -> actor.handle(message));
-                if(success) {
-                    getChannel().basicAck(envelope.getDeliveryTag(), false);
-                }
-                else {
-                    getChannel().basicReject(envelope.getDeliveryTag(), false);
-                }
-            } catch (Throwable t) {
-                log.error("Error processing message...", t);
-                if (isExceptionIgnorable(t)) {
-                    log.warn("Acked message due to exception: ", t);
-                    getChannel().basicAck(envelope.getDeliveryTag(), false);
-                }
-                else {
-                    getChannel().basicReject(envelope.getDeliveryTag(), false);
-                }
-            }
-        }
-
-    }
-
     public final void publishWithDelay(Message message, long delayMilliseconds) throws Exception {
-        log.info("Publishing message to exchange with delay: {}", delayMilliseconds);
-        if (!config.isDelayed()) {
-            log.warn("Publishing delayed message to non-delayed queue queue:{}", queueName);
-        }
-
-        if (config.getDelayType() == DelayType.TTL) {
-            publishChannel.basicPublish(ttlExchange(config),
-                    queueName,
-                    new AMQP.BasicProperties.Builder()
-                            .expiration(String.valueOf(delayMilliseconds))
-                            .deliveryMode(2)
-                            .build(),
-                    mapper().writeValueAsBytes(message));
-        } else {
-            publish(message, new AMQP.BasicProperties.Builder()
-                    .headers(Collections.singletonMap("x-delay", delayMilliseconds))
-                    .deliveryMode(2)
-                    .build());
-        }
+        actorImpl.publishWithDelay(message, delayMilliseconds);
     }
 
     public final void publish(Message message) throws Exception {
@@ -144,100 +55,18 @@ public abstract class BaseActor<Message> implements Managed {
     }
 
     public final void publish(Message message, AMQP.BasicProperties properties) throws Exception {
-        publishChannel.basicPublish(config.getExchange(), queueName, properties, mapper().writeValueAsBytes(message));
+        actorImpl.publish(message, properties);
     }
 
     @Override
     public void start() throws Exception {
-        final String exchange = config.getExchange();
-        final String dlx = config.getExchange() + "_SIDELINE";
-        if(config.isDelayed()) {
-            ensureDelayedExchange(exchange);
-        }
-        else {
-            ensureExchange(exchange);
-        }
-        ensureExchange(dlx);
+        actorImpl.start();
 
-        this.publishChannel = connection.newChannel();
-        connection.ensure(queueName + "_SIDELINE", queueName, dlx);
-        connection.ensure(queueName, config.getExchange(), connection.rmqOpts(dlx));
-        if (config.getDelayType() == DelayType.TTL) {
-            connection.ensure(ttlQueue(queueName), queueName, ttlExchange(config), connection.rmqOpts(exchange));
-        }
-        for (int i = 1; i <= config.getConcurrency(); i++) {
-            Channel consumeChannel = connection.newChannel();
-            final Handler handler = new Handler(consumeChannel,
-                    mapper, clazz, droppedExceptionTypes, prefetchCount, this);
-            final String tag = consumeChannel.basicConsume(queueName, false, handler);
-            handler.setTag(tag);
-            handlers.add(handler);
-            log.info("Started consumer {} of type {}", i, name);
-        }
-
-    }
-
-    private void ensureExchange(String exchange) throws IOException {
-        connection.channel().exchangeDeclare(
-                exchange,
-                "direct",
-                true,
-                false,
-                ImmutableMap.<String, Object>builder()
-                .put("x-ha-policy", "all")
-                .put("ha-mode", "all")
-                .build());
-    }
-
-    private void ensureDelayedExchange(String exchange) throws IOException {
-        if (config.getDelayType() == DelayType.TTL){
-            ensureExchange(ttlExchange(config));
-        } else {
-            connection.channel().exchangeDeclare(
-                    exchange,
-                    "x-delayed-message",
-                    true,
-                    false,
-                    ImmutableMap.<String, Object>builder()
-                            .put("x-ha-policy", "all")
-                            .put("ha-mode", "all")
-                            .put("x-delayed-type", "direct")
-                            .build());
-        }
-    }
-
-    private String ttlExchange(ActorConfig actorConfig) {
-        return String.format("%s_TTL", actorConfig.getExchange());
-    }
-
-    private String ttlQueue(String queueName) {
-        return String.format("%s_TTL", queueName);
     }
 
     @Override
     public void stop() throws Exception {
-        try {
-            publishChannel.close();
-        } catch (Exception e) {
-            log.error(String.format("Error closing publisher:%s" , name), e);
-        }
-        handlers.forEach(handler -> {
-            try {
-                final Channel channel = handler.getChannel();
-                channel.basicCancel(handler.getTag());
-                channel.close();
-            } catch (Exception e) {
-                log.error(String.format("Error cancelling consumer: %s", handler.getTag()), e);
-            }
-        });
-    }
-
-    protected final RMQConnection connection() throws Exception {
-        return connection;
-    }
-
-    protected final ObjectMapper mapper() {
-        return mapper;
+        actorImpl.start();
     }
 
 }
