+++ src/main/java/io/dropwizard/actors/actor/MessageHandlingFunction.java
@@ -1,0 +1,9 @@
+package io.dropwizard.actors.actor;
+
+/**
+ *
+ */
+@FunctionalInterface
+public interface MessageHandlingFunction<T,R> {
+    R apply(T param) throws Exception;
+}
